+++ apps/ui/src/types/community-actions-enum.ts
@@ -1,0 +1,5 @@
+enum MANAGER_OPTIONS {
+  EDIT = 'Edit',
+}
+
+export { MANAGER_OPTIONS };
