+++ apps/ui/src/utils/communities/cardMenuConsts.ts
@@ -1,0 +1,8 @@
+import { MANAGER_OPTIONS } from '../../types/community-actions-enum';
+
+const managerOptions: string[] = [MANAGER_OPTIONS.EDIT];
+const userOptions: string[] = [];
+
+const ITEM_HEIGHT = 48;
+
+export { managerOptions, userOptions, ITEM_HEIGHT };
